# application/services/custom_services/infra/src/swagger/swagger.yaml
---
openapi: 3.0.0
info:
   version: 1.3.0
   title: inframanager open api specification
   description: infra API
   termsOfService: http://api_url/terms/
   contact:
      name: ContactName
      email: hello@wolox.co
      url: https://www.wolox.com.ar/
   license:
      name: Apache 2.0
      url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
   - url: http://localhost:8011
     description: infra
tags:
- name: infra
- name: infracomponent
paths:
   "/infra/{id}":
      delete:
         tags:
         -  infra
         description: GpDelete
         parameters:
         -  in: "path"
            description: GpDelete
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  infra
         description: GpGetEntityById
         parameters:
         -  in: "path"
            description: GpGetEntityById
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/infra/get/search":
      get:
         tags:
         -  infra
         description: GpSearch
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/infra/get/update":
      put:
         tags:
         -  infra
         description: GpSearchForUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/infra"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/infra":
      put:
         tags:
         -  infra
         description: GpUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/infra"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  infra
         description: GpGetAllValues
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      post:
         tags:
         -  infra
         description: GpCreate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/infra"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/infra/userid/created_by":
      get:
         tags:
         -  infra
         description: GpGetNounCreatedBy
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
components:
   schemas:
      infra:
         type: object
         properties:
            providertype:
               type: "string"
               description: Description
            istemplate:
               type: "boolean"
               description: Description
            provider:
               type: "string"
               description: Description
            projectid:
               type: "string"
               description: Description
            iacframeworks:
               type: "string"
               description: Description
            iacversions:
               type: "string"
               description: Description
            components:
               type: "string"
               description: Description
      infracomponent:
         type: object
         properties:
            provider:
               type: "string"
               description: Description
            name:
               type: "string"
               description: Description
            scriptid:
               type: "string"
               description: UUID
            scriptlocation:
               type: "string"
               description: Description
            imagename:
               type: "string"
               description: Description